Day 2: feature engineering & imaging pipeline
- Add src/data/preprocessor.py: LapPreprocessor with lap filtering, IQR labeling, signal normalization
- Add src/features/imaging.py: TelemetryImager converts telemetry to (5, 200) heatmap images
- Add notebooks/day2_feature_engineering.ipynb: end-to-end pipeline from raw laps to saved dataset
- Dataset: 1023 clean laps, 61 anomalous (6%), saved to data/processed/

+++ src/features/imaging.py
@@ -3,19 +3,18 @@ import numpy as np
 class TelemetryImager:
     def __init__(self, n_bins=200):
         self.n_bins = n_bins
-    
 
     def lap_to_image(self, telemetry, preprocessor):
         signals = ['Speed', 'Throttle', 'Brake', 'nGear', 'RPM']
-    
+
         grid = np.linspace(telemetry['Distance'].min(), telemetry['Distance'].max(), self.n_bins)
-    
+
         rows = []
         for signal in signals:
-            normalized = preprocessor.normalize_signal(telemetry[signal])
+            normalized = preprocessor.normalize_signal(telemetry[signal].astype(float))
             interpolated = np.interp(grid, telemetry['Distance'], normalized)
             rows.append(interpolated)
-    
+
         return np.stack(rows)  # shape: (5, n_bins)
